Prepare for pip packaging
