+++ bench_matvec_kernel.py
@@ -3,13 +3,12 @@ import os
 import time
 
 import numpy as np
-import spqr_cuda
 import torch
 from scipy.stats import gmean
 
-from spqr import FeatureFlags, QuantizedLinear
 
-import tests.test_util as test_util
+from inference_lib.src.inference import QuantizedLinear, FeatureFlags
+from inference_lib.src.spqr.inference_kernels.cuda_kernel import call_spqr_mul_timer, call_spqr_torch_mul_timer
 
 
 def spqr_mul_timer(spqr_device: QuantizedLinear, x, feature_flag: FeatureFlags, num_runs):
@@ -18,7 +17,7 @@ def spqr_mul_timer(spqr_device: QuantizedLinear, x, feature_flag: FeatureFlags, 
 
     for i in range(num_runs):
         y = torch.zeros_like(y)
-        spqr_cuda.spqr_mul_timer(
+        call_spqr_mul_timer(
             spqr_device.m,
             spqr_device.n,
             spqr_device.bits,
@@ -37,7 +36,7 @@ def spqr_mul_timer(spqr_device: QuantizedLinear, x, feature_flag: FeatureFlags, 
     return y, runs
 
 
-def torch_mul_timer(deq_w, x, num_runs):
+def torch_mul_timer_runs(deq_w, x, num_runs):
     if len(deq_w.shape) == 1:
         n = x.shape[0]
         m = deq_w.shape[0] // n
@@ -52,7 +51,7 @@ def torch_mul_timer(deq_w, x, num_runs):
 
     for i in range(num_runs):
         y = torch.zeros_like(y)
-        spqr_cuda.torch_mul_fp16(m, n, deq_w, x, y, runs[i])
+        call_spqr_torch_mul_timer(deq_w, x, y, runs[i])
 
     return y, runs
 
@@ -111,7 +110,7 @@ if __name__ == "__main__":
             for n in [4096, 11008]:
                 d = torch.zeros((m, n), dtype=torch.float16, device=device)
                 x = torch.zeros(n, dtype=torch.float16, device=device)
-                y, dense_runs = torch_mul_timer(d, x, NUM_RUNS)
+                y, dense_runs = torch_mul_timer_runs(d, x, NUM_RUNS)
                 this_algorithm = dense_runs[WARMUP:].min()
                 torch_runs[(m, n)] = this_algorithm
                 torch.cuda.empty_cache()
@@ -176,7 +175,11 @@ if __name__ == "__main__":
                 spqr_module.to(device=device)
                 spqr_module_device = spqr_module
 
-                x_fp32 = test_util.generate_x_fp32(n)
+                def generate_x_fp32(n, upper_bound=3):
+                    x_fp32 = ((torch.rand(n) - 0.5) * 4 * upper_bound).int()
+                    return x_fp32.float()
+
+                x_fp32 = generate_x_fp32(n)
                 x_fp16_device = x_fp32.cuda(device=device).half()
 
                 deq_w_device = deq_w.to(device).half().flatten()
