--- build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-MAX_JOBS=16 python3 setup.py install
-
