+++ inference_demo.py
@@ -75,7 +75,6 @@ class InferenceDemo:
                     self.config = AutoConfig.from_pretrained(
                         pretrained_model_path, torchscript=self.torchscript, return_dict=True
                     )
-                    self.config.max_position_embeddings = 4096
 
                     self.model = AutoModelForCausalLM.from_pretrained(
                         pretrained_model_name_or_path=pretrained_model_path,
@@ -173,7 +172,7 @@ if __name__ == "__main__":
     m = Mode(args.execution_mode)
 
     with torch.no_grad():
-        model = InferenceDemo(args.pretrained_model_path, args.compressed_model_path, m)
+        model = InferenceDemo(args.pretrained_model_path, args.compressed_model_path, m, backend='inductor')
         text = "The recipe for banana bread is "  # input()
         s = time.time()
         generated_text, timings_s = model.generate(text, max_new_tokens=128)
