+++ inference_lib/setup.cfg
@@ -1,0 +1,46 @@
+[metadata]
+name = spqr_quant
+version = 0.1.0
+author = Elvir Crncevic
+author_email = elvircrn@gmail.com
+description = Efficiently run models quantized with SPQR
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/Vahe1994/SPQR
+project_urls =
+    Bug Tracker = https://github.com/Vahe1994/SPQR/issues
+classifiers =
+    Development Status :: 4 - Beta
+    Intended Audience :: Developers
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
+    Topic :: Scientific/Engineering
+    Topic :: Scientific/Engineering :: Mathematics
+    Topic :: Scientific/Engineering :: Artificial Intelligence
+    Topic :: Software Development
+    Topic :: Software Development :: Libraries
+    Topic :: Software Development :: Libraries :: Python Modules
+
+[options]
+package_dir =
+    = src
+packages = find:
+include_package_data = True
+python_requires = >=3.11
+install_requires =
+    torch>=2.4.0
+    transformers>=4.44.0
+    accelerate>=0.27.0
+[options.extras_require]
+gpu =
+    ninja
+dev =
+    pytest==6.2.5
+    pytest-forked
+    pytest-asyncio==0.16.0
+    black==22.3.0
+    isort==5.13.2
+[options.packages.find]
+where = src
