+++ inference_lib/src/__init__.py
@@ -1,0 +1,1 @@
+from .spqr.inference_kernels import *
